+++ src/param.h
@@ -162,7 +162,7 @@ public:
     vector<string> VcfSampleID; // size=total sample #
     vector<string> InputSampleID; //size = ni_total
     vector<pair<long long int, double> > UcharTable;
-    vector<uchar> SNPmean;
+    vector<double> SNPmean;
     
 	bool mode_silence;
 	int a_mode;				//analysis mode, 1/2/3/4 for Frequentist tests
